Fix Implicit narrowing conversion in compound (#259)

+++ ecaudit/src/main/java/com/ericsson/bss/cassandra/ecaudit/auth/SchemaHelper.java
@@ -53,7 +53,7 @@ class SchemaHelper
     {
         LOG.info("Waiting for schema to align in cluster");
 
-        int delayMillis = 0;
+        long delayMillis = 0;
         while (areLiveNodesWaitingOnSchemaUpdate())
         {
             if (delayMillis >= schemaAlignmentDelayMillis)
